[refactor] login to main activity
week1에서 작성한 기능을 week2,3의 fragment로 설정하고, login 화면에서 이어지도록 설정함
