+++ app/src/main/java/org/android/go/sopt/login/LoginActivity.kt
@@ -6,6 +6,7 @@ import android.content.Intent
 import android.content.SharedPreferences
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
 import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
 import androidx.activity.result.ActivityResultLauncher
@@ -13,6 +14,7 @@ import androidx.activity.result.contract.ActivityResultContracts
 import com.google.android.material.snackbar.Snackbar
 import org.android.go.sopt.R
 import org.android.go.sopt.databinding.ActivityLoginBinding
+import org.android.go.sopt.main.MainActivity
 
 class LoginActivity : AppCompatActivity() {
     private lateinit var binding: ActivityLoginBinding
@@ -68,7 +70,7 @@ class LoginActivity : AppCompatActivity() {
         // 로그인 조건 확인
         if (idEntered == id && pwEntered == pw) {
 
-            val intent = Intent(this, MyPageActivity::class.java).apply {
+            val intent = Intent(this, MainActivity::class.java).apply {
                 putExtra("id", id)
                 putExtra("pw", pw)
                 putExtra("name", name)
@@ -119,7 +121,7 @@ class LoginActivity : AppCompatActivity() {
         val idShared = sharedPreferences.getString("id", null)
         val pwShared = sharedPreferences.getString("pw", null)
         if (idShared != null && pwShared != null) {
-            val intent = Intent(this, MyPageActivity::class.java).apply {
+            val intent = Intent(this, MainActivity::class.java).apply {
                 putExtra("id", idShared)
                 putExtra("pw", pwShared)
             }
