+++ app/src/main/res/values/strings.xml
@@ -28,4 +28,6 @@
     <string name="menu_item_account">ACCOUNT</string>
     <string name="playlist_title">상호의 플레이리스트</string>
     <string name="btn_delete">DELETE</string>
+    <string name="main_upper_text">김상호님 ,</string>
+    <string name="main_lower_text">방문을 환영합니다</string>
 </resources>
